requirements.txt
poke-env
python-dotenv
torch
numpy
matplotlib